tests: Move aarch64_acpi to integration_aarch64_acpi.rs
Extract mod aarch64_acpi from integration.rs into a
standalone test binary integration_aarch64_acpi.rs.
This follows the pattern established by
integration_cvm.rs.
